Move service methods to the new bootstrapper class
