+++ packages/framework/src/Foundation/Services/LoadYamlConfiguration.php
@@ -5,7 +5,15 @@ declare(strict_types=1);
 namespace Hyde\Foundation\Services;
 
 use Hyde\Framework\Services\YamlConfigurationService;
+use Hyde\Hyde;
+use Illuminate\Support\Facades\Config;
 use LaravelZero\Framework\Application;
+use Symfony\Component\Yaml\Yaml;
+
+use function array_merge;
+use function file_exists;
+use function file_get_contents;
+use function is_array;
 
 /**
  * @internal
@@ -20,8 +28,38 @@ class LoadYamlConfiguration
      */
     public function bootstrap(Application $app): void
     {
-        if (YamlConfigurationService::hasFile()) {
-            YamlConfigurationService::boot();
+        if (static::hasFile()) {
+            static::boot();
         }
+    }
+
+    public static function boot(): void
+    {
+        if (static::hasFile()) {
+            Config::set('site', array_merge(
+                Config::get('site', []),
+                static::getYaml()
+            ));
+        }
+    }
+
+    public static function hasFile(): bool
+    {
+        return file_exists(Hyde::path('hyde.yml'))
+            || file_exists(Hyde::path('hyde.yaml'));
+    }
+
+    protected static function getFile(): string
+    {
+        return file_exists(Hyde::path('hyde.yml'))
+            ? Hyde::path('hyde.yml')
+            : Hyde::path('hyde.yaml');
+    }
+
+    protected static function getYaml(): array
+    {
+        $yaml = Yaml::parse(file_get_contents(static::getFile()));
+
+        return is_array($yaml) ? $yaml : [];
     }
 }
