+++ packages/framework/src/Framework/Services/YamlConfigurationService.php
@@ -19,33 +19,5 @@ use Symfony\Component\Yaml\Yaml;
  */
 class YamlConfigurationService
 {
-    public static function boot(): void
-    {
-        if (static::hasFile()) {
-            Config::set('site', array_merge(
-                Config::get('site', []),
-                static::getYaml()
-            ));
-        }
-    }
-
-    public static function hasFile(): bool
-    {
-        return file_exists(Hyde::path('hyde.yml'))
-            || file_exists(Hyde::path('hyde.yaml'));
-    }
-
-    protected static function getFile(): string
-    {
-        return file_exists(Hyde::path('hyde.yml'))
-            ? Hyde::path('hyde.yml')
-            : Hyde::path('hyde.yaml');
-    }
-
-    protected static function getYaml(): array
-    {
-        $yaml = Yaml::parse(file_get_contents(static::getFile()));
-
-        return is_array($yaml) ? $yaml : [];
-    }
+    //
 }
